Title: 정수 덮기, Time: 8ms, Memory: 0MB, Status: Passed - CodeTree

+++ 240915/정수 덮기/essence-cover.cpp
@@ -7,7 +7,7 @@ int main() {
     int n, l;
     cin >> n >> l;
 
-    int term[n];
+    float term[n];
 
     // 구간 중심점 입력
     for (int i = 0; i < n; i++) {
@@ -17,7 +17,7 @@ int main() {
     // 정렬
     sort(term, term+n);
     
-    int cover = term[0] - 0.5 + l;
+    float cover = term[0] - 0.5 + l;
     int cnt = 1;
     
     for (int i = 1; i < n; i++) {
